requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.2
boto3==1.15.12
botocore==1.18.12
certifi==2020.4.5.1
chardet==3.0.4
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.0.5
django-bootstrap4==2.2.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-heroku==0.3.1
django-registration==3.1.1
django-request==1.5.5
django-storages==1.10.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
heroku==0.1.4
idna==2.9
jmespath==0.10.0
Markdown==3.2.1
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.2.0
